enhance slider pointer handling
