+++ src/forms/slider/hook/index.ts
@@ -1,0 +1,1 @@
+export * from './use-slider'
