+++ src/forms/slider/index.ts
@@ -1,1 +1,2 @@
 export * from './slider'
+export * from './hook'
