+++ todo.md
@@ -13,7 +13,7 @@
 - [x] only add data attributes on essential elements in `checkbox`, unwrap `dataAttr` memo
 - [x] simplify `radio-group` 's classes, remove unnecessary wrapper elements, let form-field component to control form title & description, and make it more semantic and accessible.
 - [x] keep tooltip panel open and add move transition when hover on another trigger which also has a tooltip, to have better user experience when switching between triggers.
-- [ ] extract logic from slider component to a separate hook, and export component level hooks (should located at same dir 's `/hook` dir ) custom implementations.
+- [x] extract logic from slider component to a separate hook, and export component level hooks (should located at same dir 's `/hook` dir ) custom implementations.
 - [ ] unify custom render function prop with prefix `render`, target should be `renderXXXX`
 - [ ] unify and correct class/style priority: top level `class` / `style` (root only) > `classes` / `styles` > component builtin classes / styles
 - [ ] refactor form / form-field to formisch, replace existing form context logic if possible
